Implement opcode 8XY5: Subtract Vy from Vx with borrow
This commit adds support for the 8XY5 opcode, which subtracts the value in register Vy from the value in register Vx. If there is a borrow (i.e., if Vx < Vy), the borrow flag (VF) is set to 1. Otherwise, VF is set to 0. This instruction is used for performing subtraction with borrow detection.

- Subtract the value in register Vy from the value in register Vx.
- Set the borrow flag according to the specification definition (i.e, if Vx > Vy)
- Store the result in register Vx.

This implementation has been tested with programs that use arithmetic subtraction with borrow, ensuring compatibility with CHIP-8 specifications.