Implement opcode 8XY7: Subtract Vx from Vy with borrow
This commit adds support for the 8XY7 opcode, which subtracts the value in register Vx from the value in register Vy. If there is a borrow (i.e., if Vy < Vx), the borrow flag (VF) is set to 1. Otherwise, VF is set to 0. This instruction is used for performing subtraction with borrow detection in reverse order compared to opcode 8XY5.

- Subtract the value in register Vx from the value in register Vy.
- Set the borrow flag (VF) if there is a borrow (i.e., if Vy < Vx).
- Store the result in register Vy.

This implementation has been tested with programs that use arithmetic subtraction with borrow in reverse order, ensuring compatibility with CHIP-8 specifications.